scripts/requirements.txt
PyPDF2
python-docx
